内存与性能优化 删掉WordBlock里面的tmp字段 优化7z读取性能 修复CodeWriter在同时读写LDC时的bug 加入缓存DirectAccessor，并且改名叫Bypass 更新native，提升bsdiff的性能 然而，你需要知道的是，更新不等于没有bug，native bsdiff依然是薛定谔的稳定，崩溃是随机的，反正我搞不懂

+++ libcpp/lz/LZ_Jni.cpp
@@ -230,7 +230,7 @@ JNIEXPORT jlong JNICALL Java_roj_archive_qz_xz_LZMA2WriterN_nWrite(JNIEnv *env, 
     size_t retVal = ((FL2_compressStream) funcptr[6]) (ctx, ob, ib);
 
     if (ib->pos == ib->size) ib->pos = ib->size = 0;
-    env->SetIntField(self, fields[2], ib->pos);
+    env->SetIntField(self, fields[2], static_cast<jint>(ib->pos));
 
     env->SetIntField(self, fields[5], static_cast<jint>(ob->pos));
     return retVal;
